+++ src/test/java/org/yorm/records/Invoice.java
@@ -1,5 +1,5 @@
 package org.yorm.records;
 
-public record Invoice(long id, long company_id) {
+public record Invoice(int id, int company_id) {
 
 }
